move addbuttons to separate files
