+++ frontend/quickcheck/src/components/projectPage/AddAreaButton.jsx
@@ -1,0 +1,88 @@
+import { React, useState } from 'react';
+import {
+  useDisclosure,
+  Modal,
+  ModalOverlay,
+  ModalContent,
+  ModalBody,
+  ModalFooter,
+  ModalCloseButton,
+  ModalHeader,
+  Button,
+  IconButton,
+} from '@chakra-ui/react';
+import { AddIcon } from '@chakra-ui/icons';
+
+import Selection from '../Selection.jsx';
+
+const areaMock = {
+  0: { id: 0, name: 'Kredit', category: 'Privat' },
+  1: { id: 1, name: 'Kunden', category: 'Business' },
+  2: { id: 2, name: 'Payments', category: 'Privat' },
+};
+
+const fetchAllAreas = () => {
+  return Object.values(areaMock);
+};
+
+export default function AddAreaButton({ onAdd }) {
+  const { isOpen, onOpen, onClose } = useDisclosure();
+  const allAreas = fetchAllAreas();
+  const [selectedArea, setSelectedArea] = useState();
+  const header = 'Add Product Area';
+
+  // TODO: replace this var with state management
+  const existingAreas = []; // mock
+
+  const getAreaFromName = (areaName) => {
+    return allAreas.filter((m) => m.name === areaName)[0];
+  };
+
+  return (
+    <>
+      <IconButton
+        icon={<AddIcon />}
+        aria-label="Add Product Area"
+        onClick={onOpen}
+        variant="primary"
+        size="lg"
+        w={16}
+      ></IconButton>
+
+      <Modal isOpen={isOpen} onClose={onClose}>
+        <ModalOverlay />
+        <ModalContent>
+          <ModalHeader color="primary">{header}</ModalHeader>
+          <ModalCloseButton />
+
+          <ModalBody px={10}>
+            <Selection
+              placeholder="Select Poduct Area..."
+              options={allAreas
+                .filter((area) => !existingAreas.includes(area.id)) // filter out areas that already exist
+                .map((e) => e.name)}
+              onChange={(e) => setSelectedArea(e.target.value)}
+            />
+          </ModalBody>
+
+          <ModalFooter>
+            <Button
+              variant="primary"
+              mr={3}
+              disabled={selectedArea === undefined}
+              onClick={(e) => {
+                onAdd(getAreaFromName(selectedArea).id);
+                onClose();
+              }}
+            >
+              Save
+            </Button>
+            <Button onClick={onClose} variant="whisper">
+              Cancel
+            </Button>
+          </ModalFooter>
+        </ModalContent>
+      </Modal>
+    </>
+  );
+}
